The accelerometer and bluetooth child process classes are moved into their own source file.
